Move config and http services into their own packages
